More robust temporary file creation
Use "open(...)" with "O_CREAT | O_EXCL" to ensure that the creation of a
temporary file does not encounter a race condition with a possible other
instance of simplearchiver creating a temporary file. Note that this
entirely depends on "open(...)" with "O_CREAT | O_EXCL" for this (as
mentioned in `man 2 open`).